Build Mach-O 64-bit executable x86_64

+++ fortune.c
@@ -1226,7 +1226,7 @@ register STRFILE	*tp;
 {
 	tp->str_numstr = 0;
 	tp->str_longlen = 0;
-	tp->str_shortlen = ~((unsigned long)0);
+	tp->str_shortlen = ~((unsigned int)0);
 }
 
 /*
